Finally execute on full mustachified vision for event.php
To wit:

  - Create a concept of Page that knows how to find and load a Mustache template for a complete page view (including shared header and footer)
  - Cut the various parts of event.php up into Pages and move them into their own dir and namespace.
  - Remove a lot of subsequently dead code.
  - Put a couple of shared funcs into pages/eventHelpers.php
  - A little bit of autoload hacking to make all this work (to be converted to PSR-12 style soon(tm))
  - Introduce an abstract EventFrame subclass of Page that knows about a couple of things shared amongst most of the event.php Pages.

Over the course of these changes event.php has gone from an XSS-infested nightmare of ~2500 lines to something that lmost kind of makes sense, at least a little bit. I hope I didn't break too much along the way.

event.php definitely the most complex of all the codebase so hopefully applying this pattern elsewhere will be easy by comparison.
