+++ gatherling/pages/AuthFailed.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gatherling\Pages;
+
+class AuthFailed extends Page {
+    public string $title = 'Event Host Control Panel';
+}
