+++ gatherling/styles/ChandraNeue/css/stylesheet.css
@@ -106,6 +106,9 @@ See: https://bluebones.net/2024/09/avoid-trailing-comma-when-emitting-lists-in-m
     content: " | "
 }
 
+.highlight td {
+    color: green;
+}
 
 span.notallowed {
     font-size: larger;
