+++ gatherling/views/authFailed.mustache
@@ -1,4 +1,10 @@
-You are not permitted to make that change. Please contact the
-event host to modify this event. If you <b>are</b> the event host,
-or feel that you should have privilege to modify this event, you
-should contact the admin via <a href="https://discord.gg/2VJ8Fa6">the Discord server</a>.<br><br>
+<div class="grid_10 suffix_1 prefix_1">
+    <div id="gatherling_main" class="box">
+        <div class="uppertitle">Host Control Panel</div>
+            <p>You are not permitted to make that change. Please contact the
+            event host to modify this event. If you <b>are</b> the event host,
+            or feel that you should have privilege to modify this event, you
+                should contact the admin via <a href="https://discord.gg/2VJ8Fa6">the Discord server</a>.</p>
+        <div class="clear"></div>
+    </div>
+</div>
