+++ gatherling/views/medalList.mustache
@@ -1,42 +1,49 @@
-<form action="event.php" method="post" enctype="multipart/form-data">
-    <input type="hidden" name="name" value="{{eventName}}">
-    <table style="border-width: 0">
-        <tr>
-            <td colspan="2">
-                <input type="hidden" name="view" value="reg">
-                <table style="border-width: 0">
-                    <tr>
-                        <td class="c" colspan="2"><b>Medals</b></td>
-                    </tr>
-                    <tr>
-                        <td colspan="2" width=200>&nbsp;</td>
-                    </tr>
-                    <tr>
-                        <td class="c"><b>Medal</b></td>
-                        <td class="c"><b>Player</b></td>
-                    </tr>
-                    {{#finalists}}
-                        <tr>
-                            <td class="c">
-                                <img alt="1st" src="{{img}}" />
-                            </td>
-                            <td class="c">
-                                {{#playerDropMenu}}
-                                    {{> dropMenu}}
-                                {{/playerDropMenu}}
-                            </td>
-                        </tr>
-                    {{/finalists}}
-                    <tr>
-                        <td>&nbsp;</td>
-                    </tr>
-                    <tr>
-                        <td colspan="2" class="c">
-                            <input class="inputbutton" type="submit" name="mode" value="Update Medals">
-                        </td>
-                    </tr>
-                </table>
-            </td>
-        </tr>
-    </table>
-</form>
+<div class="grid_10 suffix_1 prefix_1">
+    <div id="gatherling_main" class="box">
+        <div class="uppertitle">Host Control Panel</div>
+        {{> controlPanel}}
+        <form action="event.php" method="post" enctype="multipart/form-data">
+            <input type="hidden" name="name" value="{{event.name}}">
+            <table style="border-width: 0">
+                <tr>
+                    <td colspan="2">
+                        <input type="hidden" name="view" value="reg">
+                        <table style="border-width: 0">
+                            <tr>
+                                <td class="c" colspan="2"><b>Medals</b></td>
+                            </tr>
+                            <tr>
+                                <td colspan="2" width=200>&nbsp;</td>
+                            </tr>
+                            <tr>
+                                <td class="c"><b>Medal</b></td>
+                                <td class="c"><b>Player</b></td>
+                            </tr>
+                            {{#finalists}}
+                                <tr>
+                                    <td class="c">
+                                        <img alt="1st" src="{{img}}" />
+                                    </td>
+                                    <td class="c">
+                                        {{#playerDropMenu}}
+                                            {{> dropMenu}}
+                                        {{/playerDropMenu}}
+                                    </td>
+                                </tr>
+                            {{/finalists}}
+                            <tr>
+                                <td>&nbsp;</td>
+                            </tr>
+                            <tr>
+                                <td colspan="2" class="c">
+                                    <input class="inputbutton" type="submit" name="mode" value="Update Medals">
+                                </td>
+                            </tr>
+                        </table>
+                    </td>
+                </tr>
+            </table>
+        </form>
+        <div class="clear"></div>
+    </div>
+</div>
