+++ gatherling/views/page.mustache
@@ -1,0 +1,3 @@
+{{> header }}
+{{{contentSafe}}}
+{{> footer}}
