+++ gatherling/views/partials/controlPanel.mustache
@@ -1,0 +1,17 @@
+<table style="border-width: 0">
+    <tr>
+        <td>&nbsp;</td>
+    </tr>
+    <tr>
+        <td class="c">
+            <ol class="inline-list pipe">
+                {{#controlPanelLinks}}
+                    <li><a href="{{href}}">{{text}}</a></li>
+                {{/controlPanelLinks}}
+            </ol>
+        </td>
+    </tr>
+    <tr>
+        <td>&nbsp;</td>
+    </tr>
+</table>
