+++ gatherling/views/partials/eventStandings.mustache
@@ -1,0 +1,57 @@
+<p></p>
+<table class="c">
+    <th class="c" colspan="8">
+        <h3>Current Standings</h3>
+    </th>
+    <tr>
+        <td colspan="8">
+            <h6>{{event.name}}</h6>
+        </td>
+    </tr>
+    <tr>
+        <td>Rank</td>
+        <td>Player</td>
+        <td>Match Points</td>
+        <td>OMW %</td>
+        <td>PGW %</td>
+        <td>OGW %</td>
+        <td>Matches Played</td>
+        <td>Byes</td>
+    </tr>
+    {{#standings}}
+        <tr {{#shouldHighlight}}class="highlight"{{/shouldHighlight}}>
+            <td>{{rank}}</td>
+            <td>{{#gameName}}{{> gameName}}{{/gameName}}</td>
+            <td>{{matchScore}}</td>
+            <td>{{oPMatch}}</td>
+            <td>{{pLGame}}</td>
+            <td>{{oPGame}}</td>
+            <td>{{matchesPlayed}}</td>
+            <td>{{byes}}</td>
+        </tr>
+    {{/standings}}
+    <tr>
+        <td colspan="8">
+            <br><b>Tiebreakers Explained</b>
+            <p></p>
+        </td>
+    </tr>
+    <tr>
+        <td colspan="8">Players with the same number of match points are ranked based on three tiebreakers scores according to DCI rules. In order, they are:</td>
+    </tr>
+    <tr>
+        <td colspan="8">OMW % is the average percentage of matches your opponents have won.</td>
+    </tr>
+    <tr>
+        <td colspan="8">PGW % is the percentage of games you have won. /td>
+    </tr>
+    <tr>
+        <td colspan="8">OGW % is the average percentage of games your opponents have won.</td>
+    </tr>
+    <tr>
+        <td colspan="8">BYEs are not included when calculating standings. For example, a player with one BYE, one win, and one loss has a match win percentage of .50 rather than .66</td>
+    </tr>
+    <tr>
+        <td colspan="8">When calculating standings, any opponent with less than a .33 win percentage is calculated as .33</td>
+    </tr>
+</table>
