+++ gatherling/views/standingsList.mustache
@@ -1,0 +1,8 @@
+<div class="grid_10 suffix_1 prefix_1">
+    <div id="gatherling_main" class="box">
+        <div class="uppertitle">Host Control Panel</div>
+        {{> controlPanel}}
+        {{#eventStandings}}{{> eventStandings}}{{/eventStandings}}
+        <div class="clear"></div>
+    </div>
+</div>
